requirements.txt
numpy
scipy
librosa
pytorch-lightning
pandas
matplotlib
scikit-learn==0.24.1
mpi4py
joblib
smplx
transforms3d
x-transformers
madmom
ftfy
axial_positional_embedding
einops
entmax
lightning-bolts
loguru
